draft(): using macro_tool
