+++ Cargo.toml
@@ -9,4 +9,5 @@ serde = { version = "=1.0.228", features = ["derive"] }
 serde_json = "=1.0.149"
 dotenvy = "=0.15.7"
 reqwest = { version = "=0.12.28", features = ["json", "rustls-tls"], default-features = false }
-rig-core = "=0.34.0"
+rig-core = { version = "=0.34.0", features = ["derive"] }
+futures = "=0.3.32"
