+++ src/tools/mod.rs
@@ -1,0 +1,14 @@
+mod args;
+mod classify_incident;
+mod context7_enrichment;
+mod parse_log;
+mod suggest_fix;
+
+pub use classify_incident::ClassifyIncident as ClassifyIncidentTool;
+pub use parse_log::ParseLog as ParseLogTool;
+pub use suggest_fix::SuggestFixTool;
+
+pub(crate) use args::{RawLogArgs, missing_raw_log_error};
+pub(crate) use context7_enrichment::{
+    Context7Resolution, Context7ResolutionError, resolve_context7_snippets,
+};
